snapcraft: include libglew based on bionic package name
As the package is now using core18 as it's base, due to
requiring CMake with certain capabilities, naturally
it requires an Ubuntu 18.04 build system to build on.

build.snapcraft.io doesn't provide a way to specify the
build environment for the base snap, in this case the
non-default core18.

For the adventurous: Building the package works inside
a Linux container (LXD, maybe systemd-nspawn), so you
could spin a build for your target device for yourself
for now.

Ubuntu 18.04 based core18 is now a hard requirement.